undo two irrelevant renames
