+++ newapp/helpers.go
@@ -1,4 +1,4 @@
-package baseapp
+package app
 
 import (
 	sdk "github.com/irisnet/irishub/types"
