+++ newapp/options.go
@@ -1,4 +1,4 @@
-package baseapp
+package app
 
 import (
 	"fmt"
