+++ newapp/protocol/router.go
@@ -1,4 +1,4 @@
-package router
+package protocol
 
 import (
 	"regexp"
