+++ newapp/protocol/v0/genesis.go
@@ -1,4 +1,4 @@
-package protocol
+package v0
 
 import (
 	"encoding/json"
