+++ newapp/protocol/v0/protocol_v0.go
@@ -1,4 +1,4 @@
-package protocol
+package v0
 
 import (
 	"fmt"
